+++ src/rich_click/rich_help_rendering.py
@@ -39,8 +39,9 @@ try:
 except ImportError:
     from rich.console import render_group as group  # type: ignore[attr-defined,no-redef]
 
-
-if CLICK_IS_BEFORE_VERSION_9X:
+if TYPE_CHECKING:
+    from rich_click.rich_command import Group as MultiCommand  # type: ignore[attr-defined]
+elif CLICK_IS_BEFORE_VERSION_9X:
     # We need to load from here to help with patching.
     from rich_click.rich_command import MultiCommand  # type: ignore[attr-defined]
 else:
@@ -282,7 +283,9 @@ def _get_option_help(
     return Columns(items)
 
 
-def _make_command_help(help_text: str, formatter: RichHelpFormatter, is_deprecated: bool) -> Union[Text, Markdown]:
+def _make_command_help(
+    help_text: str, formatter: RichHelpFormatter, is_deprecated: Union[bool, str]
+) -> Union[Text, Markdown]:
     """
     Build cli help text for a click group command.
     That is, when calling help on groups with multiple subcommands
@@ -309,8 +312,11 @@ def _make_command_help(help_text: str, formatter: RichHelpFormatter, is_deprecat
         paragraphs[0] = paragraphs[0].replace("\b\n", "")
     help_text = paragraphs[0].strip()
     if is_deprecated:
-        # TODO: Format the deprecation text.
-        help_text = f"{formatter.config.deprecated_string}{help_text}"
+        # todo: handle within config in 1.9.0?
+        if isinstance(is_deprecated, str):
+            help_text = f"{formatter.config.deprecated_with_reason_string.format(is_deprecated)}{help_text}"
+        else:
+            help_text = f"{formatter.config.deprecated_string}{help_text}"
     renderable = _make_rich_rext(help_text, formatter.config.style_option_help, formatter)
     return renderable
 
@@ -612,7 +618,7 @@ def get_rich_options(
 
 
 def get_rich_commands(
-    obj: MultiCommand,  # type: ignore[valid-type]
+    obj: MultiCommand,
     ctx: click.Context,
     formatter: RichHelpFormatter,
 ) -> None:
@@ -623,7 +629,7 @@ def get_rich_commands(
     # Look through COMMAND_GROUPS for this command
     # stick anything unmatched into a default group at the end
     cmd_groups = _resolve_groups(ctx=ctx, groups=formatter.config.command_groups, group_attribute="commands")
-    for command in obj.list_commands(ctx):  # type: ignore[attr-defined]
+    for command in obj.list_commands(ctx):
         for cmd_group in cmd_groups:
             if command in cmd_group.get("commands", []):
                 break
@@ -666,9 +672,9 @@ def get_rich_commands(
         )
         for command in cmd_group.get("commands", []):
             # Skip if command does not exist
-            if command not in obj.list_commands(ctx):  # type: ignore[attr-defined]
+            if command not in obj.list_commands(ctx):
                 continue
-            cmd = obj.get_command(ctx, command)  # type: ignore[attr-defined]
+            cmd = obj.get_command(ctx, command)
             if TYPE_CHECKING:
                 assert cmd is not None
             if cmd.hidden: