+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## Version 1.8.9 (2025-05-19)
+
+Click 8.2 support:
+
+- Fix deprecation warning in Click 8.2. [[#239](https://github.com/ewels/rich-click/pull/239)] ([@finsberg](https://github.com/finsberg))
+- Fix typing incompatibilities with Click 8.2. [[#240](https://github.com/ewels/rich-click/pull/240), [#242](https://github.com/ewels/rich-click/pull/242)] ([@finsberg](https://github.com/finsberg))
+- Fixed `no_args_is_help=True` with Click 8.2: [[#241](https://github.com/ewels/rich-click/issues/241)]
+- Added Click 8.2's support for `Parameter.deprecated: str | bool` [[#242](https://github.com/ewels/rich-click/pull/242)]
+
 ## Version 1.8.8 (2025-03-09)
 
 - Make text wrap instead of using ellipses for overflowing metavars in options tables.